use @sanity/client in connector
