+++ package.json
@@ -48,6 +48,10 @@
       "limit": "10 KB"
     }
   ],
+  "dependencies": {
+    "@nacelle/client-js-sdk": "^3.1.1",
+    "@sanity/client": "^2.1.4"
+  },
   "devDependencies": {
     "@size-limit/preset-small-lib": "^4.9.1",
     "husky": "^4.3.8",
@@ -55,8 +59,5 @@
     "tsdx": "^0.14.1",
     "tslib": "^2.1.0",
     "typescript": "^4.1.3"
-  },
-  "dependencies": {
-    "@nacelle/client-js-sdk": "^3.1.1"
   }
 }
